+++ landing/src/hooks/gql/gql.ts
@@ -17,7 +17,7 @@ import * as types from './graphql';
 const documents = {
     "\n  query getCapacitySpeedConfig($category: Enum!, $entityType: u32!) {\n    s0EternumCapacityConfigModels(where: {category: $category }) {\n        edges{\n          node {\n           weight_gram\n          }\n        }\n      }\n      s0EternumSpeedConfigModels(where: {entity_type: $entityType }) {\n        edges{\n          node {\n           sec_per_km\n          }\n        }\n      }\n  }\n": types.GetCapacitySpeedConfigDocument,
     "\n  query getEternumOwnerRealmIds($accountAddress: ContractAddress!) {\n    s0EternumOwnerModels(where: { address: $accountAddress }) {\n      edges {\n        node {\n          address\n          entity_id\n          entity {\n            models {\n              __typename\n              ... on s0_eternum_Realm {\n                realm_id\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetEternumOwnerRealmIdsDocument,
-    "\n  query getEternumEntityOwner($entityOwnerIds: [u32!]!) {\n    s0EternumEntityOwnerModels(where: { entity_owner_idIN: $entityOwnerIds }) {\n      edges {\n        node {\n          entity_id\n          entity_owner_id\n          entity {\n            models {\n              __typename\n              ... on s0_eternum_OwnedResourcesTracker {\n                resource_types\n              }\n              ... on s0_eternum_Position {\n                x\n                y\n              }\n              ... on s0_eternum_ArrivalTime {\n                arrives_at\n              }\n              ... on s0_eternum_Weight {\n                value\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetEternumEntityOwnerDocument,
+    "\n  query getEternumEntityOwner($entityOwnerIds: [u32!]!) {\n    s0EternumEntityOwnerModels(where: { entity_owner_idIN: $entityOwnerIds}, limit: 200) {\n      edges {\n        node {\n          entity_id\n          entity_owner_id\n          entity {\n            models {\n              __typename\n              ... on s0_eternum_OwnedResourcesTracker {\n                resource_types\n              }\n              ... on s0_eternum_Position {\n                x\n                y\n              }\n              ... on s0_eternum_ArrivalTime {\n                arrives_at\n              }\n              ... on s0_eternum_Weight {\n                value\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetEternumEntityOwnerDocument,
     "\n  query getAccountTokens($accountAddress: String!) {\n    tokenBalances(accountAddress: $accountAddress, limit: 8000) {\n      edges {\n        node {\n          tokenMetadata {\n            __typename\n            ... on ERC721__Token {\n              tokenId\n              metadataDescription\n              imagePath\n              contractAddress\n              metadata\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetAccountTokensDocument,
     "\n  query getERC721Mints {\n    tokenTransfers(accountAddress: \"0x0\", limit: 8000) {\n      edges {\n        node {\n          tokenMetadata {\n            __typename\n            ... on ERC721__Token {\n              tokenId\n              metadataDescription\n              imagePath\n              contractAddress\n              metadata\n            }\n          }\n        }\n      }\n    }\n  }\n": types.GetErc721MintsDocument,
     "\n  query getEntityPosition($entityIds: [u32!]!) {\n    s0EternumPositionModels(where: { entity_idIN: $entityIds }) {\n      edges {\n        node {\n          x\n          y\n          entity_id\n          entity {\n            __typename\n          }\n        }\n      }\n    }\n  }\n": types.GetEntityPositionDocument,
@@ -35,7 +35,7 @@ export function graphql(source: "\n  query getEternumOwnerRealmIds($accountAddre
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query getEternumEntityOwner($entityOwnerIds: [u32!]!) {\n    s0EternumEntityOwnerModels(where: { entity_owner_idIN: $entityOwnerIds }) {\n      edges {\n        node {\n          entity_id\n          entity_owner_id\n          entity {\n            models {\n              __typename\n              ... on s0_eternum_OwnedResourcesTracker {\n                resource_types\n              }\n              ... on s0_eternum_Position {\n                x\n                y\n              }\n              ... on s0_eternum_ArrivalTime {\n                arrives_at\n              }\n              ... on s0_eternum_Weight {\n                value\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n"): typeof import('./graphql').GetEternumEntityOwnerDocument;
+export function graphql(source: "\n  query getEternumEntityOwner($entityOwnerIds: [u32!]!) {\n    s0EternumEntityOwnerModels(where: { entity_owner_idIN: $entityOwnerIds}, limit: 200) {\n      edges {\n        node {\n          entity_id\n          entity_owner_id\n          entity {\n            models {\n              __typename\n              ... on s0_eternum_OwnedResourcesTracker {\n                resource_types\n              }\n              ... on s0_eternum_Position {\n                x\n                y\n              }\n              ... on s0_eternum_ArrivalTime {\n                arrives_at\n              }\n              ... on s0_eternum_Weight {\n                value\n              }\n            }\n          }\n        }\n      }\n    }\n  }\n"): typeof import('./graphql').GetEternumEntityOwnerDocument;
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
